+++ proton-crypto-account/src/keys/user_keys.rs
@@ -1,16 +1,17 @@
-use std::ops::Deref;
+use std::{borrow::Cow, ops::Deref};
 
 use futures::future::join_all;
+use zeroize::Zeroizing;
 
 use super::{ArmoredPrivateKey, KeyId, LockedKey, UnlockResult};
 use crate::{
     crypto::generate_locked_pgp_key,
-    errors::{AccountCryptoError, KeyError},
+    errors::{AccountCryptoError, KeyError, KeySelectionError},
     salts::KeySecret,
 };
 use proton_crypto::crypto::{
-    AsPublicKeyRef, DataEncoding, KeyGeneratorAlgorithm, PGPProviderAsync, PGPProviderSync,
-    PrivateKey, PublicKey,
+    AccessKeyInfo, AsPublicKeyRef, DataEncoding, KeyGeneratorAlgorithm, PGPProviderAsync,
+    PGPProviderSync, PrivateKey, PublicKey,
 };
 use serde::{Deserialize, Serialize};
 
@@ -74,6 +75,30 @@ impl<Provider: PGPProviderSync> Clone for UnlockedUserKeys<Provider> {
     }
 }
 
+impl<Provider: PGPProviderSync> IntoIterator for UnlockedUserKeys<Provider> {
+    type Item = UnlockedUserKey<Provider>;
+    type IntoIter = std::vec::IntoIter<UnlockedUserKey<Provider>>;
+    fn into_iter(self) -> Self::IntoIter {
+        self.0.into_iter()
+    }
+}
+
+impl<'a, Provider: PGPProviderSync> IntoIterator for &'a UnlockedUserKeys<Provider> {
+    type Item = &'a UnlockedUserKey<Provider>;
+    type IntoIter = std::slice::Iter<'a, UnlockedUserKey<Provider>>;
+    fn into_iter(self) -> Self::IntoIter {
+        self.0.iter()
+    }
+}
+
+impl<'a, Provider: PGPProviderSync> IntoIterator for &'a mut UnlockedUserKeys<Provider> {
+    type Item = &'a mut UnlockedUserKey<Provider>;
+    type IntoIter = std::slice::IterMut<'a, UnlockedUserKey<Provider>>;
+    fn into_iter(self) -> Self::IntoIter {
+        self.0.iter_mut()
+    }
+}
+
 impl<Provider: PGPProviderSync> UnlockedUserKeys<Provider> {
     /// Retrieves the primary user key for encryption and signing operations
     /// for the user who owns these keys.
@@ -82,6 +107,72 @@ impl<Provider: PGPProviderSync> UnlockedUserKeys<Provider> {
         // For now we treat the first key in the list as primary.
         // - This might change with key transparency in place.
         self.0.first()
+    }
+
+    /// Transforms the unlocked user keys into a user key selector.
+    ///
+    /// The selector can be use to seclect keys for `OpenPGP` operations.
+    pub fn into_selector<'a>(self) -> UserKeySelector<'a, Provider>
+    where
+        Self: 'a,
+        Provider: 'a,
+    {
+        UserKeySelector::new(self)
+    }
+
+    /// Creates a user key selector from the unlocked user keys.
+    ///
+    /// The selector can be use to seclect keys for `OpenPGP` operations.
+    pub fn selector(&self) -> UserKeySelector<'_, Provider> {
+        UserKeySelector::new_with_ref(self)
+    }
+
+    /// Serializes the unlocked user keys for recovery into a single binary blob.
+    ///
+    /// WARNING: This function encodes secret user material and should be used with caution.
+    pub fn serialize_to_recovery_blob(
+        &self,
+        pgp_provider: &Provider,
+    ) -> Result<Zeroizing<Vec<u8>>, AccountCryptoError> {
+        let mut recovery_blob = Zeroizing::new(Vec::new());
+        for key in &self.0 {
+            let key_bytes = pgp_provider
+                .private_key_export_unlocked(&key.private_key, DataEncoding::Bytes)
+                .map_err(AccountCryptoError::KeyExport)?;
+            recovery_blob.extend_from_slice(key_bytes.as_ref());
+        }
+        Ok(recovery_blob)
+    }
+
+    /// Deserializes the unlocked user keys from a recovery blob.
+    ///
+    /// The returned keys have fingerprints as key ids instead of the BE keys ids.
+    /// To lock them with a new secret use [`LocalUserKey::relock_user_key`].
+    pub fn deserialize_from_recovery_blob(
+        pgp_provider: &Provider,
+        recovery_blob: impl AsRef<[u8]>,
+    ) -> Result<Self, AccountCryptoError> {
+        let unlocked_keys = pgp_provider
+            .private_keys_import_unlocked(recovery_blob.as_ref())
+            .map_err(AccountCryptoError::KeyImport)?;
+        let mut unlocked_user_keys = Vec::with_capacity(unlocked_keys.len());
+        for unlocked_key in unlocked_keys {
+            let public_key = pgp_provider
+                .private_key_to_public_key(&unlocked_key)
+                .map_err(AccountCryptoError::KeyImport)?;
+            let id = unlocked_key.key_fingerprint();
+            unlocked_user_keys.push(DecryptedUserKey {
+                id: KeyId::from(id.to_string()),
+                private_key: unlocked_key,
+                public_key,
+            });
+        }
+        Ok(Self(unlocked_user_keys))
+    }
+
+    /// Returns the number of unlocked user keys.
+    pub fn num_keys(&self) -> usize {
+        self.0.len()
     }
 }
 
@@ -205,6 +296,78 @@ impl UserKeys {
             unlocked_keys: decrypted_user_keys,
             failed: failed_keys,
         }
+    }
+}
+
+/// Key selector for the unlocked user keys of an account.
+pub struct UserKeySelector<'a, P: PGPProviderSync> {
+    user_keys: Cow<'a, UnlockedUserKeys<P>>,
+}
+
+impl<'a, P: PGPProviderSync> UserKeySelector<'a, P> {
+    pub fn new(user_keys: UnlockedUserKeys<P>) -> Self {
+        Self {
+            user_keys: Cow::Owned(user_keys),
+        }
+    }
+
+    pub fn new_with_ref(user_keys: &'a UnlockedUserKeys<P>) -> Self {
+        Self {
+            user_keys: Cow::Borrowed(user_keys),
+        }
+    }
+
+    /// Returns the primary user key of this account.
+    pub fn primary(&self) -> Result<&UnlockedUserKey<P>, KeySelectionError> {
+        self.user_keys
+            .primary()
+            .ok_or(KeySelectionError::NoPrimaryUserKey)
+    }
+
+    /// Returns the public key for encryption of this account.
+    pub fn for_encryption(&self) -> Result<&P::PublicKey, KeySelectionError> {
+        self.primary().map(AsPublicKeyRef::as_public_key)
+    }
+
+    /// Returns the private key for signing of this account.
+    pub fn for_signing(&self) -> Result<&P::PrivateKey, KeySelectionError> {
+        self.primary().map(AsRef::as_ref)
+    }
+
+    /// Returns the user keys for signature verification of this account.
+    #[must_use]
+    pub fn for_signature_verification(&self) -> &[UnlockedUserKey<P>] {
+        &self.user_keys
+    }
+
+    /// Returns the user keys for decryption of this account.
+    #[must_use]
+    pub fn for_decryption(&self) -> &[UnlockedUserKey<P>] {
+        &self.user_keys
+    }
+
+    /// Transform into the raw unlocked user keys.
+    ///
+    /// Only use this function if you absolutely need to access the raw unlocked user keys.
+    #[must_use]
+    pub fn into_raw_keys(self) -> UnlockedUserKeys<P> {
+        self.user_keys.into_owned()
+    }
+}
+
+impl<'a, P: PGPProviderSync> From<UnlockedUserKeys<P>> for UserKeySelector<'a, P>
+where
+    Self: 'a,
+    P: 'a,
+{
+    fn from(user_keys: UnlockedUserKeys<P>) -> Self {
+        Self::new(user_keys)
+    }
+}
+
+impl<'a, P: PGPProviderSync> From<&'a UnlockedUserKeys<P>> for UserKeySelector<'a, P> {
+    fn from(user_keys: &'a UnlockedUserKeys<P>) -> Self {
+        Self::new_with_ref(user_keys)
     }
 }