+++ proton-crypto/src/crypto/mod.rs
@@ -36,7 +36,7 @@ mod armor;
 pub use armor::*;
 
 /// `PGPProvider` provides access to an `OpenPGP` implementation.
-pub trait PGPProvider: Send + Sync {
+pub trait PGPProvider: Send + Sync + 'static {
     /// An `OpenPGP` session key type.
     type SessionKey: SessionKey;
 
@@ -222,6 +222,14 @@ pub trait PGPProviderSync: PGPProvider {
         private_key: impl AsRef<[u8]>,
         encoding: DataEncoding,
     ) -> crate::Result<Self::PrivateKey>;
+
+    /// Imports multiple unlocked PGP private keys from a single binary blob.
+    ///
+    /// The binary blob is a concatenation of the unlocked private keys.
+    fn private_keys_import_unlocked(
+        &self,
+        private_keys: impl AsRef<[u8]>,
+    ) -> crate::Result<Vec<Self::PrivateKey>>;
 
     /// Export the PGP private key without locking it.
     ///