+++ proton-crypto/src/rust/mod.rs
@@ -1,4 +1,4 @@
-#[cfg(feature = "rustpgp")]
+#[cfg(all(feature = "rustpgp", not(forcego)))]
 pub mod pgp;
 
 pub mod srp;